refactor(federated): split engine options and stamp tests under the 500-line cap
Two files this wave pushed over the hard cap. Both moves are mechanical — the
relocated text is byte-identical to what it replaced (verified by diffing the
extracted blocks against HEAD), and no behavior or test logic changes.

internal/federated/engine.go 508 -> 451: EngineOption, WithPlanCache,
WithLogger, WithCorruptPathRetention, WithFlushVisibilityGrace and the
defaultCorruptPathRetention const move to the new
internal/federated/engine_options.go (67 lines). The seam-specific
WithParquetSource deliberately stays in parquet_source.go with the interface it
injects — options were already split that way, so this follows the existing
grain rather than centralizing for its own sake. "time" and "go.uber.org/zap"
left engine.go with their only users.

internal/federated/parquet_schema_validation_test.go 504 -> 356: the
stamp-feeder cases (stampCols plus the six TestValidate*Stamp* tests) move to
parquet_schema_validation_stamps_test.go, 188 -> 342, joining the stamp-keying
cases they share a subject with. "maps" follows them. Test count across the
pair is unchanged at 24 (19+5 before, 13+11 after).
