+++ internal/federated/parquet_schema_validation_test.go
@@ -3,7 +3,6 @@ package federated
 import (
 	"context"
 	"fmt"
-	"maps"
 	"strings"
 	"testing"
 	"time"
@@ -339,153 +338,6 @@ func TestValidateCachedPathContributesColumnsWithoutReprobe(t *testing.T) {
 	require.True(t, complete)
 	require.Contains(t, union, "score")
 	require.Len(t, exec.probes, probesAfterFirst, "cached path must not re-probe")
-}
-
-// stampCols renders the fixture DESCRIBE pairs as a manifest column stamp
-// (name → DuckDB type), the shape FileEntry.Columns carries (#256).
-func stampCols(pairs [][2]string) map[string]string {
-	stamp := map[string]string{}
-	for _, p := range pairs {
-		stamp[p[0]] = p[1]
-	}
-	return stamp
-}
-
-// A stamped path that satisfies the invariant is validated with ZERO DuckDB
-// probes; its columns feed the union and complete stays true (#256).
-func TestValidateStampedPathSkipsProbe(t *testing.T) {
-	// The executor has no canned answer for this path, so any probe would both
-	// be counted and fail — a silent fallback cannot sneak past these asserts.
-	exec := &scriptedDescribeExecutor{}
-	v := newParquetSchemaValidator()
-	stamps := map[string]map[string]string{
-		"s3://b/1/a.parquet": stampCols(append(buildValidSystemCols(), [2]string{"score", "INTEGER"})),
-	}
-
-	union, complete, err := v.Validate(context.Background(), exec, []string{"s3://b/1/a.parquet"}, stamps)
-	require.NoError(t, err)
-	require.Empty(t, exec.probes, "a valid stamp must answer the invariant with zero DuckDB probes")
-	require.True(t, complete, "a stamped path contributed its columns, so the union is complete")
-	require.Contains(t, union, "score")
-	require.Equal(t, "UUID", union["row_id"])
-
-	// The stamp feeds the same cache a probe would, and the cache owns its
-	// copy of BOTH halves of the entry: mutating the caller's map afterwards
-	// must reach neither the cached columns nor the cached cache key.
-	//
-	// The re-validation passes a fresh map holding the ORIGINAL stamp content,
-	// which is what makes this a proof rather than a coincidence. If the entry
-	// had aliased the caller's map, the delete below would have stripped
-	// "score" from the recorded stamp too, the stamp-keyed lookup would miss,
-	// and the fallback probe — which this executor cannot answer — would fail
-	// the assertions instead of quietly passing.
-	original := maps.Clone(stamps["s3://b/1/a.parquet"])
-	delete(stamps["s3://b/1/a.parquet"], "score")
-	union, complete, err = v.Validate(context.Background(), exec,
-		[]string{"s3://b/1/a.parquet"}, map[string]map[string]string{"s3://b/1/a.parquet": original})
-	require.NoError(t, err)
-	require.Empty(t, exec.probes, "an unchanged stamp keeps the path cached")
-	require.True(t, complete)
-	require.Contains(t, union, "score", "the cache holds a clone, not the caller's map")
-}
-
-// A stamp violating the invariant must NOT fail the query by itself: the probe
-// runs, and a healthy footer wins. Stamps only short-circuit success.
-func TestValidateBadStampFallsBackToProbe(t *testing.T) {
-	exec := &scriptedDescribeExecutor{cols: map[string][][2]string{
-		"a.parquet": append(buildValidSystemCols(), [2]string{"probed_only", "VARCHAR"}),
-	}}
-	v := newParquetSchemaValidator()
-	stamps := map[string]map[string]string{ // stale/corrupt stamp: row_id missing
-		"s3://b/1/a.parquet": {"changed_at": "BIGINT", "deleted_at": "BIGINT", "stamp_only": "VARCHAR"},
-	}
-
-	union, complete, err := v.Validate(context.Background(), exec, []string{"s3://b/1/a.parquet"}, stamps)
-	require.NoError(t, err, "a corrupt manifest stamp must not fail a healthy object")
-	require.Len(t, exec.probes, 1, "the rejected stamp falls through to exactly one footer probe")
-	require.True(t, complete)
-	require.Contains(t, union, "probed_only", "the union is built from the PROBED footer")
-	require.NotContains(t, union, "stamp_only", "the rejected stamp must not contribute columns")
-	require.Equal(t, "UUID", union["row_id"])
-}
-
-// Bad stamp + probe confirming the violation keeps today's loud failure.
-func TestValidateBadStampProbeConfirmsViolation(t *testing.T) {
-	exec := &scriptedDescribeExecutor{cols: map[string][][2]string{
-		"wrong.parquet": {{"wrong_col", "INTEGER"}},
-	}}
-	v := newParquetSchemaValidator()
-	stamps := map[string]map[string]string{"s3://b/1/wrong.parquet": {"wrong_col": "INTEGER"}}
-
-	_, _, err := v.Validate(context.Background(), exec, []string{"s3://b/1/wrong.parquet"}, stamps)
-	require.ErrorIs(t, err, ErrFederatedReadFailed)
-	require.Contains(t, err.Error(), "wrong.parquet")
-	require.Contains(t, err.Error(), "row_id")
-	require.Len(t, exec.probes, 1, "byte truth authors the failure — the stamp only declined to skip it")
-}
-
-// Mixed set: stamped paths skip probing, unstamped paths probe as today.
-func TestValidateMixedStampedAndUnstamped(t *testing.T) {
-	exec := &scriptedDescribeExecutor{cols: map[string][][2]string{
-		"unstamped.parquet": append(buildValidSystemCols(), [2]string{"probed", "VARCHAR"}),
-	}}
-	v := newParquetSchemaValidator()
-	stamps := map[string]map[string]string{
-		"s3://b/1/stamped.parquet": stampCols(append(buildValidSystemCols(), [2]string{"stamped_col", "INTEGER"})),
-	}
-
-	union, complete, err := v.Validate(context.Background(), exec,
-		[]string{"s3://b/1/stamped.parquet", "s3://b/1/unstamped.parquet"}, stamps)
-	require.NoError(t, err)
-	require.Len(t, exec.probes, 1, "only the unstamped path reaches DuckDB")
-	require.Contains(t, exec.probes[0], "unstamped.parquet")
-	require.True(t, complete)
-	require.Contains(t, union, "stamped_col", "the stamped path contributes its columns")
-	require.Contains(t, union, "probed", "the probed path contributes its columns")
-}
-
-// nil stamps map preserves today's behavior byte-for-byte (regression pin for
-// every pre-#256 call site).
-func TestValidateNilStampsUnchanged(t *testing.T) {
-	exec := &scriptedDescribeExecutor{cols: map[string][][2]string{
-		"good.parquet": append(buildValidSystemCols(), [2]string{"score", "INTEGER"}),
-	}}
-	v := newParquetSchemaValidator()
-
-	union, complete, err := v.Validate(context.Background(), exec, []string{"s3://b/1/good.parquet"}, nil)
-	require.NoError(t, err)
-	require.Len(t, exec.probes, 1, "without a stamp the footer probe still runs")
-	require.True(t, complete)
-	require.Contains(t, union, "score")
-
-	_, _, err = v.Validate(context.Background(), exec, []string{"s3://b/1/good.parquet"}, nil)
-	require.NoError(t, err)
-	require.Len(t, exec.probes, 1, "the probe result is still cached")
-
-	bad := &scriptedDescribeExecutor{cols: map[string][][2]string{"wrong.parquet": {{"wrong_col", "INTEGER"}}}}
-	_, _, err = newParquetSchemaValidator().Validate(context.Background(), bad,
-		[]string{"s3://b/1/wrong.parquet"}, nil)
-	require.ErrorIs(t, err, ErrFederatedReadFailed, "violations stay loud when no stamp is supplied")
-}
-
-// The glob branch forwards stamps to its expanded matches: a stamped match
-// skips its probe while an unstamped sibling is probed as today (#256).
-func TestValidateGlobMatchesConsumeStamps(t *testing.T) {
-	exec := &scriptedDescribeExecutor{
-		globs: map[string][]string{"1/*.parquet": {"s3://b/1/stamped.parquet", "s3://b/1/plain.parquet"}},
-		cols:  map[string][][2]string{"plain.parquet": buildValidSystemCols()},
-	}
-	v := newParquetSchemaValidator()
-	stamps := map[string]map[string]string{
-		"s3://b/1/stamped.parquet": stampCols(append(buildValidSystemCols(), [2]string{"stamped_col", "INTEGER"})),
-	}
-
-	union, complete, err := v.Validate(context.Background(), exec, []string{"s3://b/1/*.parquet"}, stamps)
-	require.NoError(t, err)
-	require.Len(t, exec.probes, 2, "the glob listing plus one DESCRIBE for the unstamped match only")
-	require.Contains(t, exec.probes[1], "plain.parquet")
-	require.True(t, complete)
-	require.Contains(t, union, "stamped_col")
 }
 
 // describeOverridingExecutor overrides the DESCRIBE answer while delegating
